+++ Frontend/src/components/outfits/OutfitDetailModal.jsx
